+++ src/main/java/seedu/address/logic/commands/DeleteCommand.java
@@ -9,7 +9,7 @@ import seedu.address.commons.core.index.Index;
 import seedu.address.logic.CommandHistory;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
-import seedu.address.model.person.Expense;
+import seedu.address.model.epiggy.Expense;
 
 /**
  * Deletes a expense identified using it's displayed index from the address book.
@@ -35,14 +35,14 @@ public class DeleteCommand extends Command {
     @Override
     public CommandResult execute(Model model, CommandHistory history) throws CommandException {
         requireNonNull(model);
-        List<Expense> lastShownList = model.getFilteredPersonList();
+        List<Expense> lastShownList = model.getFilteredExpenseList();
 
         if (targetIndex.getZeroBased() >= lastShownList.size()) {
             throw new CommandException(Messages.MESSAGE_INVALID_PERSON_DISPLAYED_INDEX);
         }
 
         Expense expenseToDelete = lastShownList.get(targetIndex.getZeroBased());
-        model.deletePerson(expenseToDelete);
+        model.deleteExpense(expenseToDelete);
         model.commitExpenseList();
         return new CommandResult(String.format(MESSAGE_DELETE_PERSON_SUCCESS, expenseToDelete));
     }
