+++ src/main/java/seedu/address/logic/commands/SelectCommand.java
@@ -9,7 +9,7 @@ import seedu.address.commons.core.index.Index;
 import seedu.address.logic.CommandHistory;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
-import seedu.address.model.person.Expense;
+import seedu.address.model.epiggy.Expense;
 
 /**
  * Selects a expense identified using it's displayed index from the address book.
@@ -36,7 +36,7 @@ public class SelectCommand extends Command {
     public CommandResult execute(Model model, CommandHistory history) throws CommandException {
         requireNonNull(model);
 
-        List<Expense> filteredExpenseList = model.getFilteredPersonList();
+        List<Expense> filteredExpenseList = model.getFilteredExpenseList();
 
         if (targetIndex.getZeroBased() >= filteredExpenseList.size()) {
             throw new CommandException(Messages.MESSAGE_INVALID_PERSON_DISPLAYED_INDEX);
